NAV-29213 Opprett felter-mappe og flere mindre fixes
Omdøp BehandlingstemaSelect til BehandlingstemaFelt.
Omdøp Behandlingskategori til Behandlingstema.
Legg til null og undefined for kategori og søknadMottattDato til backend.
