+++ src/frontend/komponenter/Saklinje/Meny/OpprettBehandling/felter/BehandlingstemaFelt.tsx
@@ -2,19 +2,19 @@ import {
     OpprettBehandlingFelt,
     type OpprettBehandlingFormValues,
 } from '@komponenter/Saklinje/Meny/OpprettBehandling/useOpprettBehandlingSkjema';
-import { BehandlingKategori } from '@typer/behandlingstema';
+import { Behandlingstema } from '@typer/behandlingstema';
 import { useController, useFormContext } from 'react-hook-form';
 
 import { Select } from '@navikt/ds-react';
 
-export function BehandlingstemaSelect() {
+export function BehandlingstemaFelt() {
     const { control } = useFormContext<OpprettBehandlingFormValues>();
     const {
         field: { value, onChange },
         fieldState: { error },
         formState: { isSubmitting },
     } = useController({
-        name: OpprettBehandlingFelt.BEHANDLINGSKATEGORI,
+        name: OpprettBehandlingFelt.BEHANDLINGSTEMA,
         control,
         rules: {
             required: 'Behandlingstema må velges.',
@@ -32,10 +32,10 @@ export function BehandlingstemaSelect() {
             <option disabled={true} value={''}>
                 Velg behandlingstema
             </option>
-            <option aria-selected={value === BehandlingKategori.NASJONAL} value={BehandlingKategori.NASJONAL}>
+            <option aria-selected={value === Behandlingstema.NASJONAL} value={Behandlingstema.NASJONAL}>
                 Nasjonal
             </option>
-            <option aria-selected={value === BehandlingKategori.EØS} value={BehandlingKategori.EØS}>
+            <option aria-selected={value === Behandlingstema.EØS} value={Behandlingstema.EØS}>
                 EØS
             </option>
         </Select>
